requirements.txt
accelerate==1.8.1
azure-ai-formrecognizer==3.3.3
azure-search-documents==11.5.2
azure-storage-blob==12.25.1
faiss-cpu==1.11.0
nltk==3.9.1
openai==1.90.0
python-dotenv==1.1.0
safetensors==0.5.3
sentence-transformers==4.1.0
torch==2.7.1
transformers==4.52.4